Changes as a result of review
