+++ .gitignore
@@ -75,3 +75,4 @@ fastlane/test_output
 # Visual studio code
 .vscode
 Example/Podfile.lock
+.qodo
