--- OptimoveSDK/Sources/Classes/EmbeddedMessaging/Models/EmbeddedMessageResponse.swift
@@ -1,8 +1,0 @@
-public typealias EmbeddedMessagesResponse = [String: EmbeddedMessagingContainer]
-
-
-public struct EmbeddedMessagingAPIResponse: Codable {
-    public let containers: [String: [EmbeddedMessage]]
-}
-
-
