--- OptimoveSDK/Sources/Classes/EmbeddedMessaging/Models/EmbeddedMessagingConfig.swift
@@ -1,5 +1,0 @@
-public struct EmbeddedMessagingConfig: Codable {
-    let region: String
-    let tenantId: Int
-    let brandId: String
-}
